feat: 코드 파일 내용 조회 기능 개발
