--- backend/src/main/java/solid/backend/CodeFile/Dto/CodeFileDelDto.java
@@ -1,12 +1,0 @@
-package solid.backend.CodeFile.Dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class CodeFileDelDto {
-    private Integer codeFileId;
-}
