+++ backend/src/main/java/solid/backend/CodeFile/dto/CodeFileDelDto.java
@@ -1,4 +1,4 @@
-package solid.backend.CodeFile.Dto;
+package solid.backend.CodeFile.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,7 +7,6 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class CodeFileUpdDto {
+public class CodeFileDelDto {
     private Integer codeFileId;
-    private String codeContent;
 }
