+++ backend/src/main/java/solid/backend/CodeFile/dto/CodeFileSaveDto.java
@@ -1,0 +1,14 @@
+package solid.backend.CodeFile.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class CodeFileSaveDto {
+    private Integer directoryId;
+    private String codeFileName;
+    private String codeContent;
+}
