+++ backend/src/main/java/solid/backend/CodeFile/dto/CodeFileUpdDto.java
@@ -1,4 +1,4 @@
-package solid.backend.CodeFile.Dto;
+package solid.backend.CodeFile.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,8 +7,7 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class CodeFileSaveDto {
-    private Integer directoryId;
-    private String codeFileName;
+public class CodeFileUpdDto {
+    private Integer codeFileId;
     private String codeContent;
 }
